chore: add express app setup
